refactor(listings) add blueprint and move animals routes: * move listings routes into animal_routes.py and use blueprint * AnimalRepository - create_new_animal 'image' field is now an empty string * token_checker wrapper moved to auth.py * get rid of unused imports and some unneccesary comments
